requirements.txt
# Core dependencies
pymysql==1.1.0
pandas==2.2.3
python-dotenv==1.0.1
plotly
mysql-connector-python
sqlalchemy
schedule
seaborn
matplotlib
numpy
